Wrapped Proposal detail stepper logic in function.
